Broke out models for user and thought. Created get post for users and authentication.
